Move fetchers to graphql. closes #34
