+++ src/test/java/dwbh/api/graphql/fetchers/VotingFetcherTests.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with DWBH.  If not, see <https://www.gnu.org/licenses/>
  */
-package dwbh.api.fetchers;
+package dwbh.api.graphql.fetchers;
 
 import static io.github.benas.randombeans.api.EnhancedRandom.random;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -31,7 +31,7 @@ import dwbh.api.domain.Voting;
 import dwbh.api.domain.input.CreateVoteInput;
 import dwbh.api.domain.input.CreateVotingInput;
 import dwbh.api.domain.input.ListVotingsGroupInput;
-import dwbh.api.fetchers.utils.FetcherTestUtils;
+import dwbh.api.graphql.fetchers.utils.FetcherTestUtils;
 import dwbh.api.services.VotingService;
 import dwbh.api.util.Result;
 import graphql.execution.DataFetcherResult;
