+++ src/test/java/dwbh/api/graphql/fetchers/utils/FetcherTestUtils.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with DWBH.  If not, see <https://www.gnu.org/licenses/>
  */
-package dwbh.api.fetchers.utils;
+package dwbh.api.graphql.fetchers.utils;
 
 import dwbh.api.domain.User;
 import dwbh.api.graphql.Context;
